+++ Tracker/Screens/Onboarding/TrackerListVC.swift
@@ -102,6 +102,8 @@ final class TrackerListViewController: UIViewController, DataStoreDelegate {
     override func viewDidLoad() {
         super.viewDidLoad()
         
+        
+        
         view.backgroundColor = UIColor(named: "YPWhite")
         configureNavigationBar()
         dateFromDatePicker = prepareDate(date: datePicker.date)
@@ -351,16 +353,5 @@ final class StatisticsViewController: UIViewController {
     override func viewDidLoad() {
         view.backgroundColor = UIColor(named: "YPWhite")
         title = "Статистика"
-    }
-}
-
-final class TabBarViewController: UITabBarController {
-    private let trackerListVC = UINavigationController(rootViewController: TrackerListViewController(store: DataStore()))
-    private let statisticsVC = UINavigationController(rootViewController: StatisticsViewController())
-    
-    override func viewDidLoad() {
-        trackerListVC.tabBarItem = UITabBarItem(title: "Трекеры", image: UIImage(named: "iconTrackers"), tag: 0)
-        statisticsVC.tabBarItem = UITabBarItem(title: "Статистика", image: UIImage(named: "iconStatistics"), tag: 1)
-        viewControllers = [trackerListVC, statisticsVC]
     }
 }
