+++ Tracker/Screens/Onboarding/View/OnboardingVC.swift
@@ -20,6 +20,8 @@ final class OnboardingViewController: UIPageViewController {
         dataSource = self
         delegate = self
         
+        print(FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first)
+        
         orderedViewControllers = [onboarding_01, onboarding_02]
         
         if let firstVC = orderedViewControllers.first {
